% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-requests.R
\name{arc_agent}
\alias{arc_agent}
\title{Set user-agent for arcgisutils}
\usage{
arc_agent(req)
}
\arguments{
\item{req}{an httr2 request}
}
\value{
an httr2 request object
}
\description{
Override the default user-agent set by httr2 to indicate that a request
came from arcgisutils.
}
\examples{
req <- httr2::request("http://example.com")
arc_agent(req)
}
