% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-group-user-list.R
\name{arc_group_users}
\alias{arc_group_users}
\title{List users in a group}
\usage{
arc_group_users(
  group,
  name = NULL,
  member_type = NULL,
  joined_before = NULL,
  joined_after = NULL,
  sort_field = c("username", "membertype", "joined"),
  sort_order = c("asc", "desc"),
  page_size = 50,
  max_pages = Inf,
  .progress = TRUE,
  host = arc_host(),
  token = arc_token()
)
}
\arguments{
\item{group}{a group ID as a scalar string or \code{PortalGroup} created via \code{\link[=arc_group]{arc_group()}}}

\item{name}{a scalar string of a user to search for.}

\item{member_type}{default \code{NULL}. If provided must be one of \code{"admin"} or \code{"member"}.}

\item{joined_before}{default \code{NULL}. A scalar date to search for users who joined before this date.}

\item{joined_after}{default \code{NULL}. A scalar date to search for users who joined after this date.}

\item{sort_field}{default \code{"username"}. The field to sort by. Must be one of \code{"username"}, \code{"membertype"}, or \code{"joined"}.}

\item{sort_order}{optional string. One of either \code{asc} or \code{desc} for ascending or descending order respectively.}

\item{page_size}{a scalar integer between 1 and 100 indicating the number of responses per page.}

\item{max_pages}{the maximum number of pages to fetch. By default fetches all pages.}

\item{.progress}{default \code{TRUE}. Whether to display a progress bar for requests.}

\item{host}{default \code{"https://www.arcgis.com"}. The host of your ArcGIS Portal.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[=auth_code]{auth_code()}}
or related function}
}
\value{
a data.frame
}
\description{
List all users in a given group.
}
\examples{
\dontrun{
set_arc_token(auth_user())
groups <- arc_user("r-bridge-docs")$groups
arc_group_users(groups$id[1])
}
}
\references{
\href{https://developers.arcgis.com/rest/users-groups-and-items/group-users-list/}{API Reference}
}
