% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing-class.R
\name{arc_job_status}
\alias{arc_job_status}
\title{Geoprocessing Job Status}
\usage{
arc_job_status(status = character(0))
}
\arguments{
\item{status}{a scalar character. Must be one of \code{"esriJobSubmitted"},
\code{"esriJobWaiting"}, \code{"esriJobExecuting"}, \code{"esriJobSucceeded"}, \code{"esriJobFailed"},
\code{"esriJobTimedOut"}, \code{"esriJobCancelling"}, or \verb{"esriJobCancelled}".}
}
\value{
an object of class \code{arc_job_status}
}
\description{
Represents the status of a geoprocessing job.
}
\examples{
arc_job_status("esriJobSubmitted")
}
\seealso{
Other geoprocessing:
\code{\link[=arc_form_params]{arc_form_params()}},
\code{\link{arc_gp_job}},
\code{\link[=gp_job_from_url]{gp_job_from_url()}},
\code{\link{gp_params}}
}
\concept{geoprocessing}
