% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-servers.R
\name{arc_portal_servers}
\alias{arc_portal_servers}
\title{List ArcGIS Enterprise Servers}
\usage{
arc_portal_servers(id = arc_portal_self(token)[["id"]], token = arc_token())
}
\arguments{
\item{id}{the portal ID. By default it fetches the \code{id} from \code{\link[=arc_portal_self]{arc_portal_self()}}.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[=auth_code]{auth_code()}}
or related function}
}
\value{
a data.frame of servers
}
\description{
The servers resource lists the ArcGIS Server sites that have been federated with the portal.
}
\examples{
\dontrun{
set_arc_token(auth_user())
arc_portal_servers()
}
}
