% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-token.R
\name{arc_token}
\alias{arc_token}
\title{Retrieve ARCGIS Token}
\usage{
arc_token()
}
\value{
If token exists, returns scalar character string; otherwise, NULL.
}
\description{
Utility function for sanitizing token
}
\details{
Looks for an environment variable called "ARCGIS_TOKEN." If this does not
exist, it returns an empty token (i.e., NULL value).
}
\examples{
Sys.setenv(ARCGIS_TOKEN = "testit")
arc_token()
Sys.unsetenv("ARCGIS_TOKEN")
arc_token()
}
