% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{arc_user}
\alias{arc_user}
\title{User Information}
\usage{
arc_user(username, host = arc_host(), token = arc_token())
}
\arguments{
\item{username}{the username to fetch. A scalar character.}

\item{host}{default \code{"https://www.arcgis.com"}. The host of your ArcGIS Portal.}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}
}
\value{
a list of class \code{PortalUser}
}
\description{
Fetch a user's metadata based on username.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
arc_user("esri_en")
\dontshow{\}) # examplesIf}
}
\seealso{
Other portal organization:
\code{\link[=arc_group]{arc_group()}}
}
\concept{portal organization}
