% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{arcgisutils-package}
\alias{arcgisutils}
\alias{arcgisutils-package}
\title{arcgisutils: ArcGIS Utility Functions}
\description{
Developer oriented utility functions designed to be used as the building blocks of R packages that work with ArcGIS Location Services. It provides functionality for authorization, Esri JSON construction and parsing, as well as other utilities pertaining to geometry and Esri type conversions. To support 'ArcGIS Pro' users, authorization can be done via 'arcgisbinding'. Installation instructions for 'arcgisbinding' can be found at \url{https://r.esri.com/r-bridge-site/arcgisbinding/installing-arcgisbinding.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R-ArcGIS/arcgisutils}
  \item \url{https://r.esri.com/arcgisutils/}
  \item Report bugs at \url{https://github.com/R-ArcGIS/arcgisutils/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

Other contributors:
\itemize{
  \item Kenneth Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID}) [contributor]
  \item Martha Bass (\href{https://orcid.org/0009-0004-0268-5426}{ORCID}) [contributor]
}

}
