% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{as_arc_url}
\alias{as_arc_url}
\alias{as_arc_service_url}
\alias{as_arc_layer_url}
\alias{as_arc_sharing_url}
\title{Convert different types of ArcGIS URLs}
\usage{
as_arc_service_url(url, token = NULL, call = rlang::caller_env())

as_arc_layer_url(url, layer = NULL, token = NULL, call = rlang::caller_env())

as_arc_sharing_url(url, call = rlang::caller_env())
}
\arguments{
\item{url}{A valid input URL.}

\item{token}{an \code{httr2_token} from one of the provided \code{auth_} functions}

\item{call}{default \code{\link[rlang:stack]{rlang::caller_env()}}. The calling environment passed to \code{detect_errors()}.}

\item{layer}{A layer ID value. Layer names are not yet supported.}
}
\value{
A URL of the relevant type.
}
\description{
A collection of utility functions that convert between different types of
URLs by using query arguments from the URL and the Sharing API and REST API
to extract item, service, or layer metadata.
\itemize{
\item \code{\link[=as_arc_service_url]{as_arc_service_url()}} converts an item or layer URL to a service URL
\item \code{\link[=as_arc_layer_url]{as_arc_layer_url()}} converts an item or service URL to a layer or table URL
\item \code{\link[=as_arc_sharing_url]{as_arc_sharing_url()}} converts an item, user, or group URL to a Sharing API URL
}
}
\examples{

item_url <- "https://www.arcgis.com/home/item.html?id=373a421a19564b0996ff30d554be964b"

as_arc_service_url(item_url)

as_arc_layer_url(item_url)

item_layer_url <- "https://www.arcgis.com/home/item.html?id=5192f89d7c55444ba3a4933e014c9eb8&sublayer=5"

as_arc_layer_url(item_layer_url)

layer_url <- "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/NPS_Fee_Boundary/FeatureServer/0"

as_arc_service_url(layer_url)


}
\seealso{
\code{\link[=arc_url_parse]{arc_url_parse()}}
}
