% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{check_esri_url}
\alias{check_esri_url}
\title{Is x an ESRI url of the specified type?}
\usage{
check_esri_url(
  x,
  type = "service",
  values = NULL,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{type}{Default "service". \code{type} must be a string matching a service URL
type ("root", "folder", "service", "feature") or a content url type ("content",
"search", "item", "group", "user", "scene", "app", "notebook", "experience"). \code{type} can
also be a service type: "MapServer", "FeatureServer", "ImageServer",
"GeoDataServer", "GeocodeServer", "GeometryServer", "GPServer",
"WFSServer", or "WFCServer".}

\item{allow_null}{Default \code{FALSE}. If \code{FALSE}, error if input object is
\code{NULL}. If \code{TRUE} and input is \code{NULL}, invisibly return \code{NULL}.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently running
function, e.g. \code{call = caller_env()}. The corresponding function
call is retrieved and mentioned in error messages as the source
of the error.

You only need to supply \code{call} when throwing a condition from a
helper function which wouldn't be relevant to mention in the
message.

Can also be \code{NULL} or a \link[rlang:topic-defuse]{defused function call} to
respectively not display any call or hard-code a code to display.

For more information about error calls, see \ifelse{html}{\link[rlang:topic-error-call]{Including function calls in error messages}}{\link[rlang:topic-error-call]{Including function calls in error messages}}.}
}
\description{
Is x an ESRI url of the specified type?
}
\keyword{internal}
