% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{check_string}
\alias{check_string}
\title{Check if x is a string}
\usage{
check_string(
  x,
  allow_empty = TRUE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{allow_empty}{Default \code{TRUE}. If \code{FALSE}, error if input object is an
empty string, \code{""}. If \code{TRUE} and input is \code{""}, invisibly return \code{NULL}.}

\item{allow_null}{Default \code{FALSE}. If \code{FALSE}, error if input object is
\code{NULL}. If \code{TRUE} and input is \code{NULL}, invisibly return \code{NULL}.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Check if x is a string
}
\keyword{internal}
