% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing.R
\name{class_form_params}
\alias{class_form_params}
\title{Form request parameters}
\usage{
class_form_params(params = class_missing)
}
\value{
an object of class \code{esri_form_params}
}
\description{
ArcGIS endpoints make extensive use of form encoded data for the body
of http requests. Form requests require that each element has a name
and is encoded as a single string.
}
\details{
The \code{esri_form_params} class provides validation of form body parameters.
It uses a named list internally to store the parameters.
}
\seealso{
Other geoprocessing: 
\code{\link{class_job_status}()},
\code{\link{gp_job}}
}
\concept{geoprocessing}
