% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing.R
\name{class_job_status}
\alias{class_job_status}
\title{Geoprocess Job Status}
\usage{
class_job_status(status = class_missing)
}
\arguments{
\item{status}{a scalar character. Must be one of \code{"esriJobSubmitted"},
\code{"esriJobWaiting"}, \code{"esriJobExecuting"}, \code{"esriJobSucceeded"}, \code{"esriJobFailed"},
\code{"esriJobTimedOut"}, \code{"esriJobCancelling"}, or \verb{"esriJobCancelled}".}
}
\value{
an object of class \code{gp_job_status}
}
\description{
Represents the status of a geoprocessing job.
}
\seealso{
Other geoprocessing: 
\code{\link{class_form_params}()},
\code{\link{gp_job}}
}
\concept{geoprocessing}
