% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-items.R
\name{delete_items}
\alias{delete_items}
\title{Delete Portal Items}
\usage{
delete_items(
  items,
  permanent = FALSE,
  ...,
  token = arc_token(),
  error_call = rlang::caller_env()
)
}
\arguments{
\item{items}{A character vector of item IDs, a \code{PortalItem}, or a list of them.}

\item{permanent}{Boolean. Skip the recycle bin. ArcGIS Online only.}

\item{...}{these dots are for future extensions and must be empty.}

\item{token}{an object of class \code{httr2_token} as generated by \code{\link[=auth_code]{auth_code()}}
or related function}

\item{error_call}{the caller environment to be used when propagating errors.}
}
\value{
A data.frame with columns \code{itemId} and \code{success}.
}
\description{
Deletes up to 100 items owned by the signed in user.
}
\examples{
\dontrun{
item <- upload_file("penguins.parquet", "penguins")
delete_items(item)
}
}
\references{
\href{https://developers.arcgis.com/rest/users-groups-and-items/delete-items/}{API Reference}
}
\seealso{
Other portal item:
\code{\link[=arc_item]{arc_item()}},
\code{\link[=arc_item_data]{arc_item_data()}},
\code{\link[=arc_related_items]{arc_related_items()}},
\code{\link{portal_enums}},
\code{\link[=share_item]{share_item()}},
\code{\link[=upload_file]{upload_file()}}
}
\concept{portal item}
