% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-requests.R
\name{fetch_layer_metadata}
\alias{fetch_layer_metadata}
\title{Retrieve metadata}
\usage{
fetch_layer_metadata(
  url,
  token = NULL,
  path = NULL,
  query = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{url}{the url of the item.}

\item{token}{an \code{httr2_token} from one of the provided \code{auth_} functions}

\item{path}{a character vector of paths to be appended to url using \code{\link[httr2:req_url_path_append]{httr2::req_url_path_append()}}}

\item{query}{a named vector or named list of query parameters to be appended to the url using \code{\link[httr2:req_url_query]{httr2::req_url_query()}}}

\item{call}{default \code{\link[rlang:caller_env]{rlang::caller_env()}}. The calling environment passed to \code{detect_errors()}.}
}
\value{
returns a list object
}
\description{
Utility functions for feature service metadata.
}
\details{
\itemize{
\item \code{fetch_layer_metadata()} given a request, fetches the metadata by setting
the query parameter \code{f=json}
}
}
\examples{
# url is broken into parts to fit within 100 characters to avoid CRAN notes
url_parts <- c(
  "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services",
  "/USA_Counties_Generalized_Boundaries/FeatureServer/0"
)

furl <- paste0(url_parts, collapse = "")
meta <- fetch_layer_metadata(furl)
head(names(meta))
}
