% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-auth.R
\name{host}
\alias{host}
\title{Determines Portal Host}
\usage{
host()
}
\value{
A scalar character, \code{"https://www.arcgis.com"} by default.
}
\description{
Returns a scalar character indicating the host to make requests to.
}
\details{
By default, the host is ArcGIS Online <\verb{https://www.arcgis.com}>. If the
environment variable \code{ARCGIS_HOST} is set, it will be returned.
}
\examples{
host()
}
