% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{is_esri_url}
\alias{is_esri_url}
\alias{is_esri_services_url}
\alias{is_esri_item_url}
\alias{is_esri_folder_url}
\title{Is x an ESRI service or item URL?}
\usage{
is_esri_services_url(x, type = NULL)

is_esri_item_url(x, type = NULL)

is_esri_folder_url(x, metadata = NULL, ..., call = caller_env())
}
\description{
A family of functions for testing objects to determine if they are a URL for
an ESRI service or item. These internal functions are used by
\code{\link[=check_esri_url]{check_esri_url()}}. Checking if a URL is a folder URL requires the item
metadata and can't be determined based on the URL pattern alone.
\itemize{
\item \code{\link[=is_esri_services_url]{is_esri_services_url()}}: Does x match the pattern of a ESRI Service URL?
\item \code{\link[=is_esri_folder_url]{is_esri_folder_url()}}: Is x a folder URL?
\item \code{\link[=is_esri_item_url]{is_esri_item_url()}}: Does x match the pattern of a ESRI Item URL?
}
}
\keyword{internal}
