% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-url.R
\name{list_esri_url}
\alias{list_esri_url}
\alias{list_esri_service_type_patterns}
\alias{list_esri_service_url_patterns}
\alias{list_esri_url_messages}
\alias{list_esri_url_types}
\title{List ESRI url types and patterns}
\usage{
list_esri_service_type_patterns()

list_esri_service_url_patterns()

list_esri_url_messages()

list_esri_url_types(type = NULL)
}
\description{
List ESRI url types and patterns

\itemize{
\item \code{\link[=list_esri_service_type_patterns]{list_esri_service_type_patterns()}} returns a vector of ESRI service types
}

\itemize{
\item \code{\link[=list_esri_service_url_patterns]{list_esri_service_url_patterns()}} returns a vector of ESRI URL patterns for data services and items
}

\itemize{
\item \code{\link[=list_esri_url_messages]{list_esri_url_messages()}} returns a vector of messages for what defines a valid URL of the specified type
}

\itemize{
\item \code{\link[=list_esri_url_messages]{list_esri_url_messages()}} returns a vector of all URL types or a specified url type
}
}
\keyword{internal}
