% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{opts}
\alias{opts}
\title{Configuration Options}
\description{
These options are used to customize behavior of the R packages within the R-ArcGIS Bridge project.
\subsection{Custom Headers: \code{arcgis.req_headers}}{

A named list that is injected into all requests in \code{arc_base_req()} via \link[httr2:req_headers]{httr2::req_headers}.

Use this to inject, for example, a \code{Referer} header.
}

\subsection{Debug Requests: \code{arcgis.req_debug}}{

When configured by a package, will print the cURL request created by the used package via \link[httr2:httr2_translate]{httr2::httr2_translate}.
}
}
\examples{
options(
  "arcgis.req_headers" = list(
    referer = "https://myportal.com",
    `X-Custom-Header` = "sneaky-value"
  )
)
# see the headers
arc_base_req("https://arcgis.com")
}
