% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-file.R
\name{upload_file}
\alias{upload_file}
\title{Upload a File to a Portal}
\usage{
upload_file(
  path,
  title,
  type = NULL,
  tags = NULL,
  description = NULL,
  snippet = NULL,
  folder = NULL,
  host = arc_host(),
  token = arc_token(),
  error_call = rlang::caller_env()
)
}
\arguments{
\item{path}{String. Path to the file.}

\item{title}{String. Title of the new item.}

\item{type}{String. Item type. Inferred from the file extension by default.}

\item{tags}{Character vector. Tags to apply to the item.}

\item{description}{String. Description of the item.}

\item{snippet}{String. Summary of the item.}

\item{folder}{String. Folder to upload into. Defaults to the root folder.}

\item{host}{default \code{"https://www.arcgis.com"}. The host of your ArcGIS Portal.}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}

\item{error_call}{the caller environment to be used when propagating errors.}
}
\value{
An object of class \code{PortalItem}.
}
\description{
Uploads a local file to a portal as a new content item.
}
\examples{
\dontrun{
path <- tempfile(fileext = ".csv")
write.csv(head(iris), path, row.names = FALSE)
upload_file(path, "Iris sample")
}
}
\references{
\href{https://developers.arcgis.com/rest/users-groups-and-items/add-item/}{API Reference}
}
\seealso{
Other portal item:
\code{\link[=arc_item]{arc_item()}},
\code{\link[=arc_item_data]{arc_item_data()}},
\code{\link[=arc_related_items]{arc_related_items()}},
\code{\link{portal_enums}},
\code{\link[=share_item]{share_item()}}
}
\concept{portal item}
