% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing-class.R
\name{arc_form_params}
\alias{arc_form_params}
\title{Form request parameters}
\usage{
arc_form_params(params = list())
}
\arguments{
\item{params}{a named list with scalar character elements}
}
\value{
an object of class \code{arc_form_params}
}
\description{
ArcGIS endpoints make extensive use of form encoded data for the body
of http requests. Form requests require that each element has a name
and is encoded as a single string—often as json.
}
\details{
The \code{arc_form_params} class provides validation of form body parameters
ensuring that each element is a scalar string. It uses a named list
internally to store the parameters.
}
\seealso{
Other geoprocessing: 
\code{\link{arc_gp_job}},
\code{\link{arc_job_status}()},
\code{\link{gp_params}}
}
\concept{geoprocessing}
