% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing-class.R
\name{arc_job_status}
\alias{arc_job_status}
\title{Geoprocessing Job Status}
\usage{
arc_job_status(status = character(0))
}
\arguments{
\item{status}{a scalar character. Must be one of \code{"esriJobSubmitted"},
\code{"esriJobWaiting"}, \code{"esriJobExecuting"}, \code{"esriJobSucceeded"}, \code{"esriJobFailed"},
\code{"esriJobTimedOut"}, \code{"esriJobCancelling"}, or \verb{"esriJobCancelled}".}
}
\value{
an object of class \code{gp_job_status}
}
\description{
Represents the status of a geoprocessing job.
}
\seealso{
Other geoprocessing: 
\code{\link{arc_form_params}()},
\code{\link{arc_gp_job}}
}
\concept{geoprocessing}
