% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharing.R
\name{arc_portal_urls}
\alias{arc_portal_urls}
\title{Organization's URLs}
\usage{
arc_portal_urls(host = arc_host(), token = arc_token())
}
\arguments{
\item{host}{default \code{"https://www.arcgis.com"}. The host of your ArcGIS Portal.}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}
}
\value{
a data.frame
}
\description{
Returns the URLs of an organizations services.
}
\details{
See \href{https://developers.arcgis.com/rest/users-groups-and-items/urls/}{API Reference} for more information.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontshow{if (curl::has_internet()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
arc_portal_urls()
\dontshow{\}) # examplesIf}
}
\seealso{
Other portal: 
\code{\link{arc_portal_resources}()},
\code{\link{arc_portal_users}()},
\code{\link{self}}
}
\concept{portal}
