% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-geometry.R
\name{determine_esri_geo_type}
\alias{determine_esri_geo_type}
\title{Determine Esri Geometry type}
\usage{
determine_esri_geo_type(x, call = caller_env())
}
\arguments{
\item{x}{an object of class \code{data.frame}, \code{sf}, \code{sfc}, or \code{sfg}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
returns a character scalar of the corresponding Esri geometry type
}
\description{
Takes an \code{sf} or \code{sfc} object and returns the appropriate Esri geometry type.
}
\details{
Alternatively, if
#' ### Geometry type mapping
\itemize{
\item \code{POINT}:  \code{esriGeometryPoint}
\item \code{MULTIPOINT}:  \code{esriGeometryMultipoint}
\item \code{LINESTRING}:  \code{esriGeometryPolyline}
\item \code{MULTILINESTRING}:  \code{esriGeometryPolyline}
\item \code{POLYGON}:  \code{esriGeometryPolygon}
\item \code{MULTIPOLYGON}:  \code{esriGeometryPolygon}
}
}
\examples{

determine_esri_geo_type(sf::st_point(c(0, 0)))
}
