% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing.R
\name{gp_job}
\alias{gp_job}
\alias{new_gp_job}
\title{Create a Geoprocessing Job}
\usage{
new_gp_job(base_url, params = list(), token = arc_token())
}
\value{
An object of class \code{gp_job}.
}
\description{
The \code{gp_job} class is used to interact with Geoprocessing Services in
ArcGIS Online and Enterprise.
}
\details{
The \code{gp_job} uses S7 classes for the job request parameters and job status
via \code{\link[=class_form_params]{class_form_params()}} and \code{\link[=class_job_status]{class_job_status()}} respectively. Importantly,
\code{class_form_params()} ensures that parameters provided to a geoprocessing
service are all character scalars.
}
\seealso{
Other geoprocessing: 
\code{\link{class_form_params}()},
\code{\link{class_job_status}()}
}
\concept{geoprocessing}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{the URL of the job service (without \verb{/submitJob})}

\item{\code{id}}{the ID of the started job. \code{NULL} \code{self$start()} has not been called.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{params}}{returns an S7 object of class \code{esri_form_params} (see \code{\link[=class_form_params]{class_form_params()}}) the list can be accessed via \code{self$params@params}.}

\item{\code{status}}{returns the status of the geoprocessing job as an S7 object of class \code{gp_job_status} (see \code{\link[=class_job_status]{class_job_status()}}) by querying the \verb{/jobs/\{job-id\}} endpoint.}

\item{\code{results}}{returns the current results of the job by querying the \verb{/jobs/\{job-id\}/results} endpoint.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gp_job-new}{\code{gp_job$new()}}
\item \href{#method-gp_job-start}{\code{gp_job$start()}}
\item \href{#method-gp_job-cancel}{\code{gp_job$cancel()}}
\item \href{#method-gp_job-clone}{\code{gp_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-new"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_job$new(base_url, params = list(), token = arc_token())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{the URL of the job service (without \verb{/submitJob})}

\item{\code{params}}{a named list where each element is a scalar character}

\item{\code{token}}{default \code{\link[=arc_token]{arc_token()}}. The token to be used with the job.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-start"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-start}{}}}
\subsection{Method \code{start()}}{
Starts the job by calling the \verb{/submitJob} endpoint. This also sets the public field \code{id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_job$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancels a job by calling the \verb{/cancel} endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_job$cancel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gp_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
