% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoprocessing-class.R
\name{gp_job_from_url}
\alias{gp_job_from_url}
\title{Create GP Job from existing URL}
\usage{
gp_job_from_url(url, token = arc_token())
}
\arguments{
\item{url}{the url of an existing geoprocessing job}

\item{token}{default \code{\link[=arc_token]{arc_token()}}. The token to be used with the job.}
}
\description{
Create GP Job from existing URL
}
\examples{
if (interactive()) {
job_url <- "https://hydro.arcgis.com/arcgis/rest/services/Tools/Hydrology/GPServer/TraceDownstream/jobs/jfde67910074649e4a567f0adbb8af870"

gp_job_from_url(
  job_url,
  token = auth_user()
)
}
}
\seealso{
Other geoprocessing: 
\code{\link{arc_form_params}()},
\code{\link{arc_gp_job}},
\code{\link{arc_job_status}()},
\code{\link{gp_params}}
}
\concept{geoprocessing}
