% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-enums.R
\name{portal_enums}
\alias{portal_enums}
\alias{ItemAccess}
\alias{GroupAccess}
\alias{RelationshipDirection}
\alias{SortOrder}
\alias{GroupRole}
\alias{ItemSortField}
\alias{RelationshipType}
\title{Portal enumerations}
\usage{
ItemAccess(value)

GroupAccess(value)

RelationshipDirection(value)

SortOrder(value)

GroupRole(value)

ItemSortField(value)

RelationshipType(value)
}
\arguments{
\item{value}{a scalar character. One of the variants listed for the class.}
}
\value{
An object inheriting from \code{\link[s7x:Enum]{s7x::Enum}}.
}
\description{
Closed value sets used by the Portal sharing API. Each is an
\code{\link[s7x:Enum]{s7x::Enum}} which validates on construction and can be cast back to a
character with \code{\link[=as.character]{as.character()}}.
}
\examples{
ItemAccess("org")
as.character(ItemAccess("public"))

RelationshipDirection("forward")

try(ItemAccess("everyone"))
}
\seealso{
Other portal item:
\code{\link[=arc_item]{arc_item()}},
\code{\link[=arc_item_data]{arc_item_data()}},
\code{\link[=arc_related_items]{arc_related_items()}},
\code{\link[=share_item]{share_item()}}
}
\concept{portal item}
