% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal-types.R
\name{portal_types}
\alias{portal_types}
\alias{item_type}
\alias{item_keyword}
\title{Portal Item Types}
\usage{
item_type(item_type = character(0))

item_keyword(keyword = character(0))
}
\description{
Every portal item has an associated item type. Each of those item types have keywords which cna be used to help narrow down search further.
}
\references{
\href{https://developers.arcgis.com/rest/users-groups-and-items/items-and-item-types}{REST API Documentation}
}
